Add rubocop and do major cleanup

+++ app/services/bucket_service.rb
@@ -7,9 +7,7 @@ class BucketService
 
   def save_image!(filename:)
     image = bucket.images.create(filename: filename)
-    unless image.valid?
-      GcsManagementService.new(bucket.id).delete_file(filename: filename)
-    end
+    GcsManagementService.new(bucket.id).delete_file(filename: filename) unless image.valid?
 
     image
   end
